Transfer contents from App to CreateTimeTable component
