+++ client/src/App.css
@@ -36,8 +36,3 @@
     transform: rotate(360deg);
   }
 }
-
-#root
-{
-  float: none;
-}
