+++ client/src/styles/CreateTimeTable.css
@@ -1,0 +1,3 @@
+#root {
+  float: none;
+}
